Fix numerical / correctness / leak findings from the audit
Verified to build; the ROS 2 node builds and runs on top.

- KannalaBrandt8::project(Vector3d): use atan2/sqrt instead of atan2f/sqrtf so the
  double-precision fisheye projection matches its analytic Jacobian.
- G2oTypes ImuCamPose::Update/UpdateW: NormalizeRotation() returns by value and its
  result was discarded, making the periodic re-orthonormalization a no-op; assign it
  back (Rwb = ..., DR = ...).
- G2oTypes VertexPose::read: resize the local Rcw/tcw/Rbc/tbc vectors before indexing
  (were indexed while empty -> OOB/UB on graph deserialization).
- Sim3Solver::ComputeSim3: guard the quaternion imaginary-part normalization against a
  near-zero norm (near-identity rotation) that produced a NaN rotation.
- Optimizer::LocalInertialBA: set the force-stop flag before optimize() so the local
  inertial BA can actually be aborted mid-run.
- Tracking post-relocalization: remove a dead block that allocated a Frame + prev Frame
  + Preintegrated and leaked all three every frame in the reset window (its consumer,
  ResetFrameIMU(), is an unimplemented stub).

+++ src/CameraModels/KannalaBrandt8.cpp
@@ -45,8 +45,8 @@ namespace ORB_SLAM3 {
 
     Eigen::Vector2d KannalaBrandt8::project(const Eigen::Vector3d &v3D) {
         const double x2_plus_y2 = v3D[0] * v3D[0] + v3D[1] * v3D[1];
-        const double theta = atan2f(sqrtf(x2_plus_y2), v3D[2]);
-        const double psi = atan2f(v3D[1], v3D[0]);
+        const double theta = atan2(sqrt(x2_plus_y2), v3D[2]);
+        const double psi = atan2(v3D[1], v3D[0]);
 
         const double theta2 = theta * theta;
         const double theta3 = theta * theta2;
